+++ R/utils.R
@@ -42,8 +42,12 @@ propagate <- function(xx, x, y, dx, dy, method=getOption("errors.propagation", "
       # propagate covariances for new object
       idx <- attr(x, "id")
       idy <- attr(y, "id")
-      for (id in union(ids(idx), ids(idy))) .covar(attr(xx, "id"), id) <-
-        rowSums(cbind(.covar(idx, id) * dx, .covar(idy, id) * dy), na.rm = TRUE)
+      for (id in ids(idx, idy)) {
+        idxid <- if (.id(idx) != id) .covar(idx, id) else errors(x)^2
+        idyid <- if (.id(idy) != id) .covar(idy, id) else errors(y)^2
+        .covar(attr(xx, "id"), id) <-
+          rowSums(cbind(idxid * dx, idyid * dy), na.rm = TRUE)
+      }
 
       # return the object
       xx